+++ CHANGELOG.md
@@ -3,12 +3,16 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [1.6.5] - 2017-09-12
+### Added
+- PASSBOLT-2383: Add + and \ to the list of allowed characters for the Resource fields: name, username and description
+
 ### Fixed
 - PASSBOLT-2371: Force the charset of the cake_sessions table in utf8
 - PASSBOLT-2325: As system administrator I shouldn't be able to execute passbolt CLI commands as root
 - PASSBOLT-2397: As system administrator I should see in the healthcheck if app/tmp content and app/webroot/img/public content are writable
 - PASSBOLT-1991: As system administrator I should see in the healthcheck if the server key can be used for encrypting/decrypting
-- PASSBOLT-2383: Add + and \ to the list of allowed characters for the Resource fields : name, username and description
 
 ## [1.6.4] - 2017-08-31
 ### Fixed
@@ -317,7 +321,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - AP: User with plugin installed
 - LU: Logged in user
 
-[Unreleased]: https://github.com/passbolt/passbolt_api/compare/v1.6.4...HEAD
+[Unreleased]: https://github.com/passbolt/passbolt_api/compare/v1.6.5...HEAD
+[1.6.5]: https://github.com/passbolt/passbolt_api/compare/v1.6.4...v1.6.5
 [1.6.4]: https://github.com/passbolt/passbolt_api/compare/v1.6.3...v1.6.4
 [1.6.3]: https://github.com/passbolt/passbolt_api/compare/v1.6.2...v1.6.3
 [1.6.2]: https://github.com/passbolt/passbolt_api/compare/v1.6.1...v1.6.2